--- a/kornia/utils/_compat.py
+++ b/kornia/utils/_compat.py
@@ -24,7 +24,6 @@ from torch import Tensor
 
 def torch_version() -> str:
     """Parse the `torch.__version__` variable and removes +cu*/cpu."""
-    # Use partition to avoid allocating a list as in split; faster for a single delimiter search.
     return torch.__version__.partition("+")[0]